Revert "Move methods to more appropriate file"
This reverts commit bab751802f80dfbf928fa3b11ed1108ffe5d0cdc.
