+++ src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTracker.cs
@@ -820,6 +820,211 @@ namespace Microsoft.CodeAnalysis
                 }
             }
 
+            private async Task<(Compilation compilationWithGeneratedFiles, CompilationTrackerGeneratorInfo generatorInfo)> AddExistingOrComputeNewGeneratorInfoAsync(
+                SolutionState solution,
+                Compilation compilationWithoutGeneratedFiles,
+                CompilationTrackerGeneratorInfo generatorInfo,
+                Compilation? compilationWithStaleGeneratedTrees,
+                CancellationToken cancellationToken)
+            {
+                if (generatorInfo.DocumentsAreFinal)
+                {
+                    // We must have ran generators before, but for some reason had to remake the compilation from
+                    // scratch. This could happen if the trees were strongly held, but the compilation was entirely
+                    // garbage collected. Just add in the trees we already have. We don't want to rerun since the
+                    // consumer of this Solution snapshot has already seen the trees and thus needs to ensure identity
+                    // of them.
+                    var compilationWithGeneratedFiles = compilationWithoutGeneratedFiles.AddSyntaxTrees(
+                        await generatorInfo.Documents.States.Values.SelectAsArrayAsync(
+                            static (state, cancellationToken) => state.GetSyntaxTreeAsync(cancellationToken), cancellationToken).ConfigureAwait(false));
+
+                    // Will reuse the generator info since we're reusing all the trees from within it.
+                    return (compilationWithGeneratedFiles, generatorInfo);
+                }
+
+                if (!this.ProjectState.SourceGenerators.Any())
+                {
+                    // We don't have any source generators.  Trivially bail out.
+                    var compilationWithGeneratedFiles = compilationWithoutGeneratedFiles;
+                    return (compilationWithGeneratedFiles, new CompilationTrackerGeneratorInfo(
+                        TextDocumentStates<SourceGeneratedDocumentState>.Empty, generatorInfo.Driver, documentsAreFinal: true));
+                }
+
+                return await ComputeNewGeneratorInfoAsync(
+                    solution,
+                    compilationWithoutGeneratedFiles,
+                    generatorInfo,
+                    compilationWithStaleGeneratedTrees,
+                    cancellationToken).ConfigureAwait(false);
+            }
+
+            private async Task<(Compilation compilationWithGeneratedFiles, CompilationTrackerGeneratorInfo generatorInfo)> ComputeNewGeneratorInfoInCurrentProcessAsync(
+                SolutionState solution,
+                Compilation compilationWithoutGeneratedFiles,
+                CompilationTrackerGeneratorInfo generatorInfo,
+                Compilation? compilationWithStaleGeneratedTrees,
+                CancellationToken cancellationToken)
+            {
+                // We have at least one source generator. If we don't already have a generator driver, we'll have to
+                // create one from scratch
+                if (generatorInfo.Driver == null)
+                {
+                    var additionalTexts = this.ProjectState.AdditionalDocumentStates.SelectAsArray(static documentState => documentState.AdditionalText);
+                    var compilationFactory = this.ProjectState.LanguageServices.GetRequiredService<ICompilationFactoryService>();
+
+                    generatorInfo = generatorInfo.WithDriver(compilationFactory.CreateGeneratorDriver(
+                        this.ProjectState.ParseOptions!,
+                        ProjectState.SourceGenerators.ToImmutableArray(),
+                        this.ProjectState.AnalyzerOptions.AnalyzerConfigOptionsProvider,
+                        additionalTexts));
+                }
+                else
+                {
+#if DEBUG
+                    // Assert that the generator driver is in sync with our additional document states; there's not a public
+                    // API to get this, but we'll reflect in DEBUG-only.
+                    var driverType = generatorInfo.Driver.GetType();
+                    var stateMember = driverType.GetField("_state", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    Contract.ThrowIfNull(stateMember);
+                    var additionalTextsMember = stateMember.FieldType.GetField("AdditionalTexts", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
+                    Contract.ThrowIfNull(additionalTextsMember);
+                    var state = stateMember.GetValue(generatorInfo.Driver);
+                    var additionalTexts = (ImmutableArray<AdditionalText>)additionalTextsMember.GetValue(state)!;
+
+                    Contract.ThrowIfFalse(additionalTexts.Length == this.ProjectState.AdditionalDocumentStates.Count);
+#endif
+                }
+
+                Contract.ThrowIfNull(generatorInfo.Driver);
+
+                // HACK HACK HACK HACK to address https://github.com/dotnet/roslyn/issues/59818. There, we were running into issues where
+                // a generator being present and consuming syntax was causing all red nodes to be processed. This was problematic when
+                // Razor design time files are also fed in, since those files tend to be quite large. The Razor design time files
+                // aren't produced via a generator, but rather via our legacy IDynamicFileInfo mechanism, so it's also a bit strange
+                // we'd even give them to other generators since that doesn't match the real compiler anyways. This simply removes
+                // all of those trees in an effort to speed things up, and also ensure the design time compilations are a bit more accurate.
+                using var _ = ArrayBuilder<SyntaxTree>.GetInstance(out var treesToRemove);
+
+                foreach (var documentState in ProjectState.DocumentStates.States)
+                {
+                    // This matches the logic in CompileTimeSolutionProvider for which documents are removed when we're
+                    // activating the generator.
+                    if (documentState.Value.Attributes.DesignTimeOnly)
+                        treesToRemove.Add(await documentState.Value.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false));
+                }
+
+                var compilationToRunGeneratorsOn = compilationWithoutGeneratedFiles.RemoveSyntaxTrees(treesToRemove);
+                // END HACK HACK HACK HACK.
+
+                generatorInfo = generatorInfo.WithDriver(generatorInfo.Driver.RunGenerators(compilationToRunGeneratorsOn, cancellationToken));
+                Contract.ThrowIfNull(generatorInfo.Driver);
+
+                var runResult = generatorInfo.Driver.GetRunResult();
+
+                var telemetryCollector = solution.Services.GetService<ISourceGeneratorTelemetryCollectorWorkspaceService>();
+                telemetryCollector?.CollectRunResult(runResult, generatorInfo.Driver.GetTimingInfo(), ProjectState);
+
+                // We may be able to reuse compilationWithStaleGeneratedTrees if the generated trees are identical. We will assign null
+                // to compilationWithStaleGeneratedTrees if we at any point realize it can't be used. We'll first check the count of trees
+                // if that changed then we absolutely can't reuse it. But if the counts match, we'll then see if each generated tree
+                // content is identical to the prior generation run; if we find a match each time, then the set of the generated trees
+                // and the prior generated trees are identical.
+                if (compilationWithStaleGeneratedTrees != null)
+                {
+                    var generatedTreeCount =
+                        runResult.Results.Sum(r => IsGeneratorRunResultToIgnore(r) ? 0 : r.GeneratedSources.Length);
+
+                    if (generatorInfo.Documents.Count != generatedTreeCount)
+                        compilationWithStaleGeneratedTrees = null;
+                }
+
+                using var generatedDocumentsBuilder = TemporaryArray<SourceGeneratedDocumentState>.Empty;
+                foreach (var generatorResult in runResult.Results)
+                {
+                    if (IsGeneratorRunResultToIgnore(generatorResult))
+                        continue;
+
+                    var generatorAnalyzerReference = this.ProjectState.GetAnalyzerReferenceForGenerator(generatorResult.Generator);
+
+                    foreach (var generatedSource in generatorResult.GeneratedSources)
+                    {
+                        var existing = FindExistingGeneratedDocumentState(
+                            generatorInfo.Documents,
+                            generatorResult.Generator,
+                            generatorAnalyzerReference,
+                            generatedSource.HintName);
+
+                        if (existing != null)
+                        {
+                            var newDocument = existing.WithUpdatedGeneratedContent(
+                                generatedSource.SourceText,
+                                this.ProjectState.ParseOptions!);
+
+                            generatedDocumentsBuilder.Add(newDocument);
+
+                            if (newDocument != existing)
+                                compilationWithStaleGeneratedTrees = null;
+                        }
+                        else
+                        {
+                            // NOTE: the use of generatedSource.SyntaxTree to fetch the path and options is OK,
+                            // since the tree is a lazy tree and that won't trigger the parse.
+                            var identity = SourceGeneratedDocumentIdentity.Generate(
+                                ProjectState.Id,
+                                generatedSource.HintName,
+                                generatorResult.Generator,
+                                generatedSource.SyntaxTree.FilePath,
+                                generatorAnalyzerReference);
+
+                            generatedDocumentsBuilder.Add(
+                                SourceGeneratedDocumentState.Create(
+                                    identity,
+                                    generatedSource.SourceText,
+                                    generatedSource.SyntaxTree.Options,
+                                    ProjectState.LanguageServices));
+
+                            // The count of trees was the same, but something didn't match up. Since we're here, at least one tree
+                            // was added, and an equal number must have been removed. Rather than trying to incrementally update
+                            // this compilation, we'll just toss this and re-add all the trees.
+                            compilationWithStaleGeneratedTrees = null;
+                        }
+                    }
+                }
+
+                // If we didn't null out this compilation, it means we can actually use it
+                if (compilationWithStaleGeneratedTrees != null)
+                    return (compilationWithStaleGeneratedTrees, generatorInfo.WithDocumentsAreFinal(true));
+
+                // We produced new documents, so time to create new state for it
+                var generatedDocuments = new TextDocumentStates<SourceGeneratedDocumentState>(generatedDocumentsBuilder.ToImmutableAndClear());
+                var compilationWithGeneratedFiles = compilationWithoutGeneratedFiles.AddSyntaxTrees(
+                    await generatedDocuments.States.Values.SelectAsArrayAsync(
+                        static (state, cancellationToken) => state.GetSyntaxTreeAsync(cancellationToken), cancellationToken).ConfigureAwait(false));
+                return (compilationWithGeneratedFiles, new CompilationTrackerGeneratorInfo(generatedDocuments, generatorInfo.Driver, documentsAreFinal: true));
+
+                static SourceGeneratedDocumentState? FindExistingGeneratedDocumentState(
+                    TextDocumentStates<SourceGeneratedDocumentState> states,
+                    ISourceGenerator generator,
+                    AnalyzerReference analyzerReference,
+                    string hintName)
+                {
+                    var generatorIdentity = SourceGeneratorIdentity.Create(generator, analyzerReference);
+
+                    foreach (var (_, state) in states.States)
+                    {
+                        if (state.Identity.Generator != generatorIdentity)
+                            continue;
+
+                        if (state.HintName != hintName)
+                            continue;
+
+                        return state;
+                    }
+
+                    return null;
+                }
+            }
+
             /// <summary>
             /// Attempts to get (without waiting) a metadata reference to a possibly in progress
             /// compilation. Only actual compilation references are returned. Could potentially
